requirements.txt
altgraph==0.16.1
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
baidu-aip==2.2.9.0
beautifulsoup4==4.5.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
Django==2.1.4
django-celery==3.2.2
django-crispy-forms==1.7.2
django-redis==4.10.0
djangorestframework==3.9.0
future==0.17.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jsonpath==0.80
kiwisolver==1.0.1
kombu==3.0.37
lxml==4.2.5
macholib==1.11
matplotlib==3.0.2
mysqlclient==1.3.14
numpy==1.15.4
pandas==0.23.4
parsel==1.5.1
pefile==2018.8.8
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.4
pymongo==3.7.2
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.3.0
pytesseract==0.2.6
python-dateutil==2.7.5
pytz==2018.7
pywin32==223.1
pywin32-ctypes==0.2.0
queuelib==1.5.0
redis==3.0.1
requests==2.21.0
Scrapy==1.5.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
zope.interface==4.6.0
